pyproject.toml
[project]
name = "devdex"
version = "0.1.0"
description = "CLI tool that scans codebases and generates deployment-ready launch kits"
requires-python = ">=3.13"
dependencies = [
    "typer>=0.15",
    "rich>=13.0",
    "questionary>=2.1",
    "pydantic>=2.0",
    "jinja2>=3.1",
    "wandb>=0.19",
    "weave>=0.51",
    "python-dotenv>=1.0",
    "gitpython>=3.1",
    "openai>=1.60",
]

[project.scripts]
devdex = "devdex.cli:app"

[project.optional-dependencies]
pipeline = ["supabase>=2.0", "mistralai>=1.0", "weave>=0.51"]
pipeline-mlx = [
    "mlx-lm[train]>=0.30",
]
pipeline-unsloth = [
    "unsloth[colab-new]",
]
pipeline-local = [
    "trl>=0.12",
    "peft>=0.13",
    "transformers>=4.45",
    "datasets>=3.0",
    "torch>=2.0",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "pytest-mock>=3.14",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["devdex"]

[tool.uv]
package = false
